gpu: Create descriptor_validation folder
One renaming:
UpdateInstrumentationBuffer to UpdateBindlessStateBuffer
