+++ BUILD.gn
@@ -171,13 +171,17 @@ vvl_sources = [
   "layers/gpu/cmd_validation/gpuav_dispatch.cpp",
   "layers/gpu/cmd_validation/gpuav_trace_rays.cpp",
   "layers/gpu/cmd_validation/gpuav_copy_buffer_to_image.cpp",
+  "layers/gpu/descriptor_validation/gpuav_descriptor_validation.h",
+  "layers/gpu/descriptor_validation/gpuav_descriptor_validation.cpp",
+  "layers/gpu/descriptor_validation/gpuav_descriptor_set.cpp",
+  "layers/gpu/descriptor_validation/gpuav_descriptor_set.h",
+  "layers/gpu/descriptor_validation/gpuav_image_layout.cpp",
   "layers/gpu/core/gpuav.h",
   "layers/gpu/core/gpuav.cpp",
   "layers/gpu/core/gpu_state_tracker.cpp",
   "layers/gpu/core/gpu_state_tracker.h",
   "layers/gpu/core/gpuav_record.cpp",
   "layers/gpu/core/gpuav_setup.cpp",
-  "layers/gpu/core/gpuav_image_layout.cpp",
   "layers/gpu/debug_printf/debug_printf.cpp",
   "layers/gpu/debug_printf/debug_printf.h",
   "layers/gpu/error_message/gpuav_error_message.cpp",
@@ -188,8 +192,6 @@ vvl_sources = [
   "layers/gpu/instrumentation/gpu_shader_instrumentor.h",
   "layers/gpu/instrumentation/gpuav_instrumentation.h",
   "layers/gpu/instrumentation/gpuav_instrumentation.cpp",
-  "layers/gpu/resources/gpuav_descriptor_set.cpp",
-  "layers/gpu/resources/gpuav_descriptor_set.h",
   "layers/gpu/resources/gpuav_subclasses.cpp",
   "layers/gpu/resources/gpuav_subclasses.h",
   "layers/gpu/resources/gpu_resources.h",
