+++ layers/gpu/core/gpuav.cpp
@@ -54,120 +54,6 @@ VkDeviceAddress Validator::GetBufferDeviceAddress(VkBuffer buffer, const Locatio
     return 0;
 }
 
-void Validator::UpdateBoundPipeline(VkCommandBuffer commandBuffer, VkPipelineBindPoint pipelineBindPoint, VkPipeline pipeline,
-                                    const Location &loc) {
-    if (!gpuav_settings.validate_descriptors) return;
-
-    auto cb_node = GetWrite<CommandBuffer>(commandBuffer);
-    if (!cb_node) {
-        InternalError(commandBuffer, loc, "Unrecognized command buffer");
-        return;
-    }
-    const auto lv_bind_point = ConvertToLvlBindPoint(pipelineBindPoint);
-    auto const &last_bound = cb_node->lastBound[lv_bind_point];
-    // Should have just been updated
-    if (!last_bound.pipeline_state) {
-        InternalError(pipeline, loc, "Unrecognized pipeline");
-        return;
-    }
-
-    // Catch if pipeline is bound before any descriptor sets
-    if (cb_node->di_input_buffer_list.empty()) {
-        return;
-    }
-
-    // Update the last vkCmdBindDescriptorSet with the new pipeline
-    auto &descriptor_set_buffers = cb_node->di_input_buffer_list.back().descriptor_set_buffers;
-
-    // If the user calls vkCmdBindDescriptorSet::firstSet to a non-zero value, these indexes don't line up
-    size_t update_index = 0;
-    for (uint32_t i = 0; i < last_bound.per_set.size(); i++) {
-        if (last_bound.per_set[i].bound_descriptor_set) {
-            auto slot = last_bound.pipeline_state->active_slots.find(i);
-            if (slot != last_bound.pipeline_state->active_slots.end()) {
-                if (update_index >= descriptor_set_buffers.size()) {
-                    // TODO - Hit crash running with Dota2, this shouldn't happen, need to look into
-                    continue;
-                }
-                descriptor_set_buffers[update_index++].binding_req = slot->second;
-            }
-        }
-    }
-}
-
-void Validator::UpdateBoundDescriptors(VkCommandBuffer commandBuffer, VkPipelineBindPoint pipelineBindPoint, const Location &loc) {
-    auto cb_node = GetWrite<CommandBuffer>(commandBuffer);
-    if (!cb_node) {
-        InternalError(commandBuffer, loc, "Unrecognized command buffer");
-        return;
-    }
-    const auto lv_bind_point = ConvertToLvlBindPoint(pipelineBindPoint);
-    auto const &last_bound = cb_node->lastBound[lv_bind_point];
-
-    uint32_t number_of_sets = static_cast<uint32_t>(last_bound.per_set.size());
-    // Figure out how much memory we need for the input block based on how many sets and bindings there are
-    // and how big each of the bindings is
-    if (number_of_sets > 0 && gpuav_settings.validate_descriptors && force_buffer_device_address) {
-        VkBufferCreateInfo buffer_info = vku::InitStructHelper();
-        assert(number_of_sets <= glsl::kDebugInputBindlessMaxDescSets);
-        buffer_info.size = sizeof(glsl::BindlessStateBuffer);
-        buffer_info.usage = VK_BUFFER_USAGE_STORAGE_BUFFER_BIT;
-        VmaAllocationCreateInfo alloc_info = {};
-        alloc_info.requiredFlags = VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT;
-        alloc_info.pool = VK_NULL_HANDLE;
-        DescBindingInfo di_buffers = {};
-
-        // Allocate buffer for device addresses of the input buffer for each descriptor set.  This is the buffer written to each
-        // draw's descriptor set.
-        VkResult result = vmaCreateBuffer(vmaAllocator, &buffer_info, &alloc_info, &di_buffers.bindless_state_buffer,
-                                          &di_buffers.bindless_state_buffer_allocation, nullptr);
-        if (result != VK_SUCCESS) {
-            InternalError(commandBuffer, loc, "Unable to allocate device memory. Device could become unstable.", true);
-            return;
-        }
-        glsl::BindlessStateBuffer *bindless_state{nullptr};
-        result =
-            vmaMapMemory(vmaAllocator, di_buffers.bindless_state_buffer_allocation, reinterpret_cast<void **>(&bindless_state));
-        if (result != VK_SUCCESS) {
-            InternalError(commandBuffer, loc, "Unable to map device memory. Device could become unstable.", true);
-            return;
-        }
-        memset(bindless_state, 0, static_cast<size_t>(buffer_info.size));
-        cb_node->current_bindless_buffer = di_buffers.bindless_state_buffer;
-
-        bindless_state->global_state = desc_heap->GetDeviceAddress();
-        for (uint32_t i = 0; i < last_bound.per_set.size(); i++) {
-            const auto &s = last_bound.per_set[i];
-            auto set = s.bound_descriptor_set;
-            if (!set) {
-                continue;
-            }
-            if (gpuav_settings.validate_descriptors) {
-                DescSetState desc_set_state;
-                desc_set_state.num = i;
-                desc_set_state.state = std::static_pointer_cast<DescriptorSet>(set);
-                bindless_state->desc_sets[i].layout_data = desc_set_state.state->GetLayoutState();
-                // The pipeline might not have been bound yet, so will need to update binding_req later
-                if (last_bound.pipeline_state) {
-                    auto slot = last_bound.pipeline_state->active_slots.find(i);
-                    if (slot != last_bound.pipeline_state->active_slots.end()) {
-                        desc_set_state.binding_req = slot->second;
-                    }
-                }
-                if (!desc_set_state.state->IsUpdateAfterBind()) {
-                    desc_set_state.gpu_state = desc_set_state.state->GetCurrentState();
-                    bindless_state->desc_sets[i].in_data = desc_set_state.gpu_state->device_addr;
-                    desc_set_state.output_state = desc_set_state.state->GetOutputState();
-                    bindless_state->desc_sets[i].out_data = desc_set_state.output_state->device_addr;
-                }
-                di_buffers.descriptor_set_buffers.emplace_back(std::move(desc_set_state));
-            }
-        }
-        cb_node->di_input_buffer_list.emplace_back(di_buffers);
-        vmaUnmapMemory(vmaAllocator, di_buffers.bindless_state_buffer_allocation);
-    }
-}
-
 bool Validator::AllocateOutputMem(gpu::DeviceMemoryBlock &output_mem, const Location &loc) {
     VkBufferCreateInfo buffer_info = vku::InitStructHelper();
     buffer_info.size = output_buffer_byte_size;
