+++ layers/gpu/core/gpuav.h
@@ -18,10 +18,9 @@
 #pragma once
 
 #include "gpu/error_message/gpuav_error_message.h"
-#include "gpu/resources/gpuav_descriptor_set.h"
+#include "gpu/descriptor_validation/gpuav_descriptor_set.h"
 #include "gpu/resources/gpu_resources.h"
 #include "gpu/instrumentation/gpu_shader_instrumentor.h"
-#include "gpu/instrumentation/gpuav_instrumentation.h"
 
 #include <unordered_map>
 #include <memory>
