+++ layers/gpu/descriptor_validation/gpuav_descriptor_validation.h
@@ -16,3 +16,11 @@
  */
 
 #pragma once
+
+#include "vma/vma.h"
+
+namespace gpuav {
+class CommandBuffer;
+
+void UpdateBindlessStateBuffer(CommandBuffer& cb_node, VmaAllocator vmaAllocator);
+}  // namespace gpuav
