+++ layers/gpu/resources/gpuav_subclasses.cpp
@@ -20,8 +20,7 @@
 #include "gpu/core/gpuav.h"
 #include "gpu/core/gpuav_constants.h"
 #include "gpu/error_message/gpuav_vuids.h"
-#include "gpu/instrumentation/gpuav_instrumentation.h"
-#include "drawdispatch/descriptor_validator.h"
+#include "gpu/descriptor_validation/gpuav_descriptor_validation.h"
 #include "gpu_shaders/gpu_error_header.h"
 
 namespace gpuav {
@@ -413,7 +412,7 @@ void CommandBuffer::ClearCmdErrorsCountsBuffer() const {
 }
 
 bool CommandBuffer::PreProcess() {
-    gpuav::UpdateInstrumentationBuffer(*this, state_.vmaAllocator);
+    gpuav::UpdateBindlessStateBuffer(*this, state_.vmaAllocator);
     const bool succeeded = UpdateBdaRangesBuffer();
     if (!succeeded) {
         return false;
@@ -483,39 +482,7 @@ void CommandBuffer::PostProcess(VkQueue queue, const Location &loc) {
     // If instrumentation found an error, skip post processing. Errors detected by instrumentation are usually
     // very serious, such as a prematurely destroyed resource and the state needed below is likely invalid.
     if (!error_found) {
-        // For each vkCmdBindDescriptorSets()...
-        // Some applications repeatedly call vkCmdBindDescriptorSets() with the same descriptor sets, avoid
-        // checking them multiple times.
-        vvl::unordered_set<VkDescriptorSet> validated_desc_sets;
-        for (auto &di_info : di_input_buffer_list) {
-            Location draw_loc(vvl::Func::vkCmdDraw);
-            // For each descriptor set ...
-            for (uint32_t i = 0; i < di_info.descriptor_set_buffers.size(); i++) {
-                auto &set = di_info.descriptor_set_buffers[i];
-                if (validated_desc_sets.count(set.state->VkHandle()) > 0) {
-                    // TODO - If you share two VkDescriptorSet across two different sets in the SPIR-V, we are not going to be
-                    // validating the 2nd instance of it
-                    continue;
-                }
-                validated_desc_sets.emplace(set.state->VkHandle());
-                assert(set.output_state);
-
-                vvl::DescriptorValidator context(state_, *this, *set.state, i, VK_NULL_HANDLE /*framebuffer*/, draw_loc);
-                const uint32_t shader_set = glsl::kDescriptorSetWrittenMask | i;
-                auto used_descs = set.output_state->UsedDescriptors(*set.state, shader_set);
-                // For each used binding ...
-                for (const auto &u : used_descs) {
-                    auto iter = set.binding_req.find(u.first);
-                    vvl::DescriptorBindingInfo binding_info;
-                    binding_info.first = u.first;
-                    while (iter != set.binding_req.end() && iter->first == u.first) {
-                        binding_info.second.emplace_back(iter->second);
-                        ++iter;
-                    }
-                    context.ValidateBinding(binding_info, u.second);
-                }
-            }
-        }
+        ValidateBindlessDescriptorSets();
     }
 
     state_.UpdateCmdBufImageLayouts(*this);
